+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.86.0] - 2023-10-04 23:17:14
+
+### Added
+
+- Quebec work premium tax credit.
+
+## [0.85.1] - 2023-10-03 23:05:56
+
+### Fixed
+
+- Disability tax credit parameter readme file.
+
 ## [0.85.0] - 2023-09-28 18:01:54
 
 ### Added
@@ -650,6 +662,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
+[0.86.0]: https://github.com/PolicyEngine/policyengine-canada/compare/0.85.1...0.86.0
+[0.85.1]: https://github.com/PolicyEngine/policyengine-canada/compare/0.85.0...0.85.1
 [0.85.0]: https://github.com/PolicyEngine/policyengine-canada/compare/0.84.0...0.85.0
 [0.84.0]: https://github.com/PolicyEngine/policyengine-canada/compare/0.83.0...0.84.0
 [0.83.0]: https://github.com/PolicyEngine/policyengine-canada/compare/0.82.0...0.83.0